Fix Clang warning

+++ Sources/HodEngine/Editor/HierachyWindow.cpp
@@ -181,7 +181,7 @@ namespace hod::editor
 		if (ImGui::BeginItemTooltip())
 		{
 			std::string_view internalStateLabel = EnumTrait::ToString(entity->GetInternalState());
-			ImGui::Text("InternalState: %.*s", internalStateLabel.size(), internalStateLabel.data());
+			ImGui::Text("InternalState: %.*s", (int)internalStateLabel.size(), internalStateLabel.data());
 			ImGui::EndTooltip();
 		}
